+++ lib/claper_web/live/attendee_live_auth.ex
@@ -1,5 +1,5 @@
 defmodule ClaperWeb.AttendeeLiveAuth do
-  import Phoenix.LiveView
+  import Phoenix.Component
 
   def on_mount(:default, _params, session, socket) do
     socket =