+++ lib/claper_web/live/event_live/show.ex
@@ -77,7 +77,7 @@ defmodule ClaperWeb.EventLive.Show do
       |> assign(:liked_posts, reacted_posts(socket, event.id, "👍"))
       |> assign(:loved_posts, reacted_posts(socket, event.id, "❤️"))
       |> assign(:loled_posts, reacted_posts(socket, event.id, "😂"))
-      |> assign(:selected_poll_opt, nil)
+      |> assign(:selected_poll_opt, [])
       |> assign(:poll_opt_saved, false)
       |> assign(:event, event)
       |> assign(:state, event.presentation_file.presentation_state)
@@ -444,50 +444,65 @@ defmodule ClaperWeb.EventLive.Show do
   def handle_event(
         "select-poll-opt",
         %{"opt" => opt},
-        socket
+        %{assigns: %{current_poll: %{multiple: true}}} = socket
       ) do
-    {:noreply, socket |> assign(:selected_poll_opt, opt)}
+        if Enum.member?(socket.assigns.selected_poll_opt, opt) do
+          {:noreply, socket |> assign(:selected_poll_opt, Enum.filter(socket.assigns.selected_poll_opt, fn x -> x != opt end))}
+        else
+          {:noreply, socket |> assign(:selected_poll_opt, [opt | socket.assigns.selected_poll_opt])}
+        end
+  end
+
+  @impl true
+  def handle_event(
+        "select-poll-opt",
+        %{"opt" => opt},
+        %{assigns: %{current_poll: %{multiple: false}}} = socket
+      ) do
+    {:noreply, socket |> assign(:selected_poll_opt, [opt])}
   end
 
   @impl true
   def handle_event(
         "vote",
-        %{"opt" => idx} = _params,
-        %{assigns: %{current_user: current_user}} = socket
+        _params,
+        %{assigns: %{current_user: current_user, selected_poll_opt: opts}} = socket
       )
       when is_map(current_user) do
-    {idx, _} = Integer.parse(idx)
-    poll_opt = Enum.at(socket.assigns.current_poll.poll_opts, idx)
 
-    case Claper.Polls.vote(
-           current_user.id,
-           socket.assigns.event.uuid,
-           poll_opt,
-           socket.assigns.current_poll.id
-         ) do
-      {:ok, poll} ->
-        {:noreply, socket |> get_current_vote(poll.id)}
+        opts = Enum.map(opts, fn opt -> Integer.parse(opt) |> elem(0) end)
+        poll_opts = Enum.map(opts, fn opt -> Enum.at(socket.assigns.current_poll.poll_opts, opt) end)
+
+        case Claper.Polls.vote(
+              current_user.id,
+              socket.assigns.event.uuid,
+              poll_opts,
+              socket.assigns.current_poll.id
+            ) do
+          {:ok, poll} ->
+            {:noreply, socket |> get_current_vote(poll.id)}
     end
   end
 
   @impl true
   def handle_event(
         "vote",
-        %{"opt" => idx} = _params,
-        %{assigns: %{attendee_identifier: attendee_identifier}} = socket
+         _params,
+         %{assigns: %{attendee_identifier: attendee_identifier, selected_poll_opt: opts}} = socket
       ) do
-    {idx, _} = Integer.parse(idx)
-    poll_opt = Enum.at(socket.assigns.current_poll.poll_opts, idx)
 
-    case Claper.Polls.vote(
-           attendee_identifier,
-           socket.assigns.event.uuid,
-           poll_opt,
-           socket.assigns.current_poll.id
-         ) do
-      {:ok, poll} ->
-        {:noreply, socket |> get_current_vote(poll.id)}
-    end
+        opts = Enum.map(opts, fn opt -> Integer.parse(opt) |> elem(0) end)
+        poll_opts = Enum.map(opts, fn opt -> Enum.at(socket.assigns.current_poll.poll_opts, opt) end)
+
+        case Claper.Polls.vote(
+              attendee_identifier,
+              socket.assigns.event.uuid,
+              poll_opts,
+              socket.assigns.current_poll.id
+            ) do
+          {:ok, poll} ->
+            {:noreply, socket |> get_current_vote(poll.id)}
+        end
   end
 
   def toggle_side_menu(js \\ %JS{}) do