+++ lib/claper_web/router.ex
@@ -7,7 +7,7 @@ defmodule ClaperWeb.Router do
     plug(:accepts, ["html"])
     plug(:fetch_session)
     plug(:fetch_live_flash)
-    plug(:put_root_layout, {ClaperWeb.LayoutView, :root})
+    plug(:put_root_layout, html: {ClaperWeb.LayoutView, :root})
     plug(:protect_from_forgery)
     plug(:put_secure_browser_headers)
     plug(:fetch_current_user)
@@ -38,7 +38,7 @@ defmodule ClaperWeb.Router do
   end
 
   live_session :user,
-    root_layout: {ClaperWeb.LayoutView, "user.html"} do
+    root_layout: {ClaperWeb.LayoutView, :user} do
     scope "/", ClaperWeb do
       pipe_through([:browser, :require_authenticated_user])