+++ lib/claper_web/views/user_registration_view.ex
@@ -1,4 +1,5 @@
 defmodule ClaperWeb.UserRegistrationView do
+  import Phoenix.Component
   use ClaperWeb, :view
 
   def render("user.json", %{user_registration: user}) do